backend/requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
cryptography==42.0.8
bcrypt==4.1.3
PyJWT==2.9.0
python-dotenv==1.0.1
APScheduler==3.10.4
